refactor: Split MCP auto-configurations into dedicated client/server modules
- Separate MCP client and server auto-configurations into dedicated modules
- Add support for both sync and async MCP clients
- Add support for STDIO, WebMVC and WebFlux transports
- Improve configuration properties organization and validation
- Add proper lifecycle management and customization support
- Add comprehensive JavaDoc documentation
