+++ auto-configurations/spring-ai-mcp-client/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,21 @@
+#
+# Copyright 2023-2024 the original author or authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+org.springframework.ai.autoconfigure.mcp.client.StdioTransportAutoConfiguration
+org.springframework.ai.autoconfigure.mcp.client.SseWebFluxTransportAutoConfiguration
+org.springframework.ai.autoconfigure.mcp.client.SseHttpClientTransportAutoConfiguration
+org.springframework.ai.autoconfigure.mcp.client.MpcClientAutoConfiguration
+
+
