+++ auto-configurations/spring-ai-mcp-client/src/test/resources/application-test.properties
@@ -1,0 +1,10 @@
+# Test MCP STDIO client configuration
+spring.ai.mcp.client.stdio.enabled=true
+spring.ai.mcp.client.stdio.version=test-version
+spring.ai.mcp.client.stdio.request-timeout=15s
+spring.ai.mcp.client.stdio.root-change-notification=false
+
+# Test server configuration
+spring.ai.mcp.client.stdio.stdio-connections.test-server.command=echo
+spring.ai.mcp.client.stdio.stdio-connections.test-server.args[0]=test
+spring.ai.mcp.client.stdio.stdio-connections.test-server.env.TEST_ENV=test-value
