+++ pom.xml
@@ -34,7 +34,12 @@
 		<module>spring-ai-bom</module>
 		<module>spring-ai-core</module>
 		<module>spring-ai-test</module>
+
 		<module>spring-ai-spring-boot-autoconfigure</module>
+
+		<module>auto-configurations/spring-ai-mcp-client</module>
+		<module>auto-configurations/spring-ai-mcp-server</module>
+
 		<module>spring-ai-retry</module>
 		<module>spring-ai-spring-boot-docker-compose</module>
 		<module>spring-ai-spring-boot-testcontainers</module>
@@ -127,9 +132,11 @@
 		<module>spring-ai-spring-boot-starters/spring-ai-starter-zhipuai</module>
 		<module>spring-ai-spring-boot-starters/spring-ai-starter-moonshot</module>
 
-		<module>spring-ai-spring-boot-starters/spring-ai-starter-mcp</module>
-		<module>spring-ai-spring-boot-starters/spring-ai-starter-mcp-webmvc</module>
-		<module>spring-ai-spring-boot-starters/spring-ai-starter-mcp-webflux</module>
+		<module>spring-ai-spring-boot-starters/spring-ai-starter-mcp-client</module>
+		<module>spring-ai-spring-boot-starters/spring-ai-starter-mcp-server</module>
+		<module>spring-ai-spring-boot-starters/spring-ai-starter-mcp-client-webflux</module>
+		<module>spring-ai-spring-boot-starters/spring-ai-starter-mcp-server-webflux</module>		
+		<module>spring-ai-spring-boot-starters/spring-ai-starter-mcp-server-webmvc</module>
 
 		<module>spring-ai-integration-tests</module>
 
@@ -887,6 +894,16 @@
 			<name>Mark Pollack</name>
 			<email>mpollack at vmware.com</email>
 			<organization>VMware</organization>
+			<organizationUrl>http://www.spring.io</organizationUrl>
+			<roles>
+				<role>lead</role>
+			</roles>
+		</developer>
+		<developer>
+			<id>tzolov</id>
+			<name>Christian Tzolov</name>
+			<email>christian tzolov at broadcom.com</email>
+			<organization>Broadcom</organization>
 			<organizationUrl>http://www.spring.io</organizationUrl>
 			<roles>
 				<role>lead</role>
