+++ spring-ai-spring-boot-autoconfigure/pom.xml
@@ -20,9 +20,6 @@
 		<developerConnection>git@github.com:spring-projects/spring-ai.git</developerConnection>
 	</scm>
 
-	<properties>
-	</properties>
-
 	<dependencies>
 
 		<dependency>
