+++ spring-ai-spring-boot-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -60,8 +60,3 @@ org.springframework.ai.autoconfigure.minimax.MiniMaxAutoConfiguration
 org.springframework.ai.autoconfigure.vertexai.embedding.VertexAiEmbeddingAutoConfiguration
 org.springframework.ai.autoconfigure.chat.memory.cassandra.CassandraChatMemoryAutoConfiguration
 org.springframework.ai.autoconfigure.vectorstore.observation.VectorStoreObservationAutoConfiguration
-org.springframework.ai.autoconfigure.mcp.server.MpcServerAutoConfiguration
-org.springframework.ai.autoconfigure.mcp.server.MpcWebMvcServerAutoConfiguration
-org.springframework.ai.autoconfigure.mcp.server.MpcWebFluxServerAutoConfiguration
-org.springframework.ai.autoconfigure.mcp.client.stdio.MpcStdioClientAutoConfiguration
-
